BatchSpawnerBase: Add background_tasks, connect_to_job feature.
This adds the possibility to start a "connect_to_job"
background task on the hub on job start,
which establishes connectivity to the actual single user server.
An example for this can be "condor_ssh_to_job" for HTCondor batch systems.

Additionally, the background tasks are monitored:
- for successful startup. The background task is given
  some time to successfully establish connectivity.
- in poll() during job runtime
and if they fail, the job is terminated.