BatchSpawnerBase: Extend connect_to_job for port forwarding approaches.
This allows to use {rport} inside the connect_to_job_cmd.
If this is done, {port} is set to a random_port chosen locally
on the Hub, and {rport} is set to the original remote port.
This is useful e.g. for SSH port forwarding.

If this is used, the {rport} of the notebook is saved into
a new class variable in case it will be needed again.